backend/requirements.txt
fastapi>=0.111.0
python-multipart>=0.0.9
uvicorn>=0.29.0
pydantic>=2.7.0,<3.0.0

langchain>=0.2.15
langchain-text-splitters>=0.2.0
langchain-community>=0.2.15
langsmith>=0.1.98
langgraph>=0.2.0

chromadb>=0.5.0
pypdf>=4.1.0
python-dotenv>=1.0.1
